refactor(uds): decompose cda-comm-uds/src/lib.rs into modules
Complete the refactoring of cda-comm-uds/src/lib.rs
by extracting data-related, query, and utility logic into dedicated modules
